Refactor Polyline and related structures for improved clarity and functionality
- Introduced BoundingBox structure for axis-aligned bounding box calculations.
- Added SignedArea and winding order checks for polylines.
- Implemented ExtractSegments to build explicit segment lists from polylines.
- Enhanced ComputeArcGeometry to handle bulge-encoded segments more effectively.
- Cleaned up and organized existing code, removing unused functions and improving readability.
- Updated Polyline to support better handling of vertex properties such as bulge and curve status.
